Refactor client script into ES modules
