+++ index.html
@@ -254,7 +254,7 @@
   </div><!-- /container -->
 
   <!-- Scripts -->
-  <script src="app.js"></script>
+    <script type="module" src="js/main.js"></script>
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
